requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0
python-dotenv==1.0.0
pydantic==2.5.0

# Google Sheets API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# OCR - Google Cloud Vision (primary) + Tesseract (fallback)
google-cloud-vision==3.5.0
pytesseract==0.3.10
Pillow==11.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.0

# PDF extraction
pdfplumber==0.10.3
PyMuPDF==1.23.8

# Optional - Database (only needed if using database features)
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2
dateparser==1.2.0

# Firecrawl
firecrawl-py

